requirements.txt
pyyaml>=6.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12